Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asgiref>=3.8.1",
    "cloudinary>=1.44.1",
    "django>=5.2.3",
    "django-cloudinary-storage>=0.3.0",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "mysqlclient>=2.2.7",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "pillow>=11.2.1",
    "pyjwt>=2.9.0",
    "pypi-json>=0.4.0",
    "python-dotenv>=1.1.1",
]
